refactor: split propagation.validation.cpp tests into four files
