+++ test/OpenSpaceToolkit/Astrodynamics/Trajectory/Validation/NRLMSIS00.validation.cpp
@@ -1,0 +1,327 @@
+/// Apache License 2.0
+
+#include <numeric>
+
+#include <OpenSpaceToolkit/Core/Containers/Array.hpp>
+#include <OpenSpaceToolkit/Core/Containers/Tuple.hpp>
+#include <OpenSpaceToolkit/Core/Types/Integer.hpp>
+#include <OpenSpaceToolkit/Core/Types/Real.hpp>
+#include <OpenSpaceToolkit/Core/Types/Shared.hpp>
+#include <OpenSpaceToolkit/Core/Types/Size.hpp>
+#include <OpenSpaceToolkit/Core/Types/String.hpp>
+
+#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Composite.hpp>
+#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Cuboid.hpp>
+#include <OpenSpaceToolkit/Mathematics/Geometry/3D/Objects/Point.hpp>
+#include <OpenSpaceToolkit/Mathematics/Objects/Vector.hpp>
+
+#include <OpenSpaceToolkit/Physics/Coordinate/Frame.hpp>
+#include <OpenSpaceToolkit/Physics/Data/Scalar.hpp>
+#include <OpenSpaceToolkit/Physics/Environment.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/CSSISpaceWeather.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/Exponential.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Atmospheric/Earth/Manager.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Gravitational/Earth.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Object.hpp>
+#include <OpenSpaceToolkit/Physics/Environment/Objects/CelestialBodies/Earth.hpp>
+#include <OpenSpaceToolkit/Physics/Time/DateTime.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Duration.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Instant.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Interval.hpp>
+#include <OpenSpaceToolkit/Physics/Time/Scale.hpp>
+#include <OpenSpaceToolkit/Physics/Units/Mass.hpp>
+
+#include <OpenSpaceToolkit/Astrodynamics/Dynamics.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Dynamics/AtmosphericDrag.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Dynamics/CentralBodyGravity.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Dynamics/PositionDerivative.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Flight/System/PropulsionSystem.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Flight/System/SatelliteSystem.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Flight/System/SatelliteSystemBuilder.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/Propagator.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesBroker.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubset.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubsets/CartesianPosition.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/CoordinatesSubsets/CartesianVelocity.hpp>
+#include <OpenSpaceToolkit/Astrodynamics/Trajectory/State/NumericalSolver.hpp>
+
+#include <Global.test.hpp>
+
+using ostk::core::ctnr::Array;
+using ostk::core::ctnr::Tuple;
+using ostk::core::types::Integer;
+using ostk::core::types::Real;
+using ostk::core::types::Shared;
+using ostk::core::types::String;
+using ostk::core::types::Size;
+
+using ostk::math::geometry::d3::objects::Composite;
+using ostk::math::geometry::d3::objects::Cuboid;
+using ostk::math::geometry::d3::objects::Point;
+using ostk::math::object::Matrix3d;
+using ostk::math::object::MatrixXd;
+using ostk::math::object::Vector3d;
+using ostk::math::object::VectorXd;
+
+using ostk::physics::coord::Frame;
+using ostk::physics::data::Scalar;
+using ostk::physics::Environment;
+using EarthAtmosphericModel = ostk::physics::environment::atmospheric::Earth;
+using EarthGravitationalModel = ostk::physics::environment::gravitational::Earth;
+using EarthMagneticModel = ostk::physics::environment::magnetic::Earth;
+using SWManager = ostk::physics::environment::atmospheric::earth::Manager;
+using ostk::physics::environment::atmospheric::earth::CSSISpaceWeather;
+using ostk::physics::environment::Object;
+using ostk::physics::environment::object::Celestial;
+using ostk::physics::environment::object::celestial::Earth;
+using ostk::physics::time::DateTime;
+using ostk::physics::time::Duration;
+using ostk::physics::time::Instant;
+using ostk::physics::time::Interval;
+using ostk::physics::time::Scale;
+using ostk::physics::units::Mass;
+
+using ostk::astro::Dynamics;
+using ostk::astro::dynamics::AtmosphericDrag;
+using ostk::astro::dynamics::CentralBodyGravity;
+using ostk::astro::dynamics::PositionDerivative;
+using ostk::astro::flight::system::PropulsionSystem;
+using ostk::astro::flight::system::SatelliteSystem;
+using ostk::astro::flight::system::SatelliteSystemBuilder;
+using ostk::astro::trajectory::State;
+using ostk::astro::trajectory::Propagator;
+using ostk::astro::trajectory::state::CoordinatesSubset;
+using ostk::astro::trajectory::state::CoordinatesBroker;
+using ostk::astro::trajectory::state::coordinatessubsets::CartesianPosition;
+using ostk::astro::trajectory::state::coordinatessubsets::CartesianVelocity;
+using ostk::astro::trajectory::state::NumericalSolver;
+
+class OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation : public ::testing::Test
+{
+   protected:
+    void SetUp() override
+    {
+        const Composite satelliteGeometry(Cuboid(
+            {0.0, 0.0, 0.0},
+            {Vector3d {1.0, 0.0, 0.0}, Vector3d {0.0, 1.0, 0.0}, Vector3d {0.0, 0.0, 1.0}},
+            {1.0, 2.0, 3.0}
+        ));
+
+        const PropulsionSystem propulsionSystem = {
+            Scalar(0.1, PropulsionSystem::thrustSIUnit),
+            Scalar(1500.0, PropulsionSystem::specificImpulseSIUnit),
+        };
+
+        this->satelliteGeometry_ = satelliteGeometry;
+        this->propulsionSystem_ = propulsionSystem;
+
+        this->satelliteSystem_ = {
+            this->satelliteDryMass_,
+            satelliteGeometry_,
+            Matrix3d::Identity(),
+            1.0,
+            2.1,
+            propulsionSystem_,
+        };
+
+        this->earthSpherical_ = std::make_shared<Celestial>(Earth::Spherical());
+        this->defaultDynamics_ = {
+            std::make_shared<PositionDerivative>(),
+            std::make_shared<CentralBodyGravity>(earthSpherical_),
+        };
+
+        this->defaultPropagator_ = {defaultNumericalSolver_, defaultDynamics_};
+    }
+
+    const NumericalSolver defaultNumericalSolver_ = {
+        NumericalSolver::LogType::NoLog,
+        NumericalSolver::StepperType::RungeKuttaFehlberg78,
+        5.0,
+        1.0e-15,
+        1.0e-15,
+    };
+
+    const Shared<const Frame> gcrfSPtr_ = Frame::GCRF();
+
+    Array<Shared<Dynamics>> defaultDynamics_ = Array<Shared<Dynamics>>::Empty();
+    const Mass satelliteDryMass_ = Mass(100.0, Mass::Unit::Kilogram);
+
+    PropulsionSystem propulsionSystem_ = PropulsionSystem::Undefined();
+    Composite satelliteGeometry_ = Composite::Undefined();
+    SatelliteSystem satelliteSystem_ = SatelliteSystem::Undefined();
+    Shared<Celestial> earthSpherical_ = nullptr;
+    Propagator defaultPropagator_ = Propagator::Undefined();
+
+    const Shared<CoordinatesBroker> dragCoordinatesBrokerSPtr_ = std::make_shared<CoordinatesBroker>(CoordinatesBroker(
+        {CartesianPosition::Default(),
+         CartesianVelocity::Default(),
+         CoordinatesSubset::Mass(),
+         CoordinatesSubset::SurfaceArea(),
+         CoordinatesSubset::DragCoefficient()}
+    ));
+};
+
+class OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Success
+    : public OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation,
+      public ::testing::WithParamInterface<Tuple<Instant, Instant, Duration, std::string>>
+{
+};
+
+INSTANTIATE_TEST_SUITE_P(
+    Input_Data_Spans_Valid,
+    OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Success,
+    testing::Values(std::make_tuple(
+        Instant::Now() - Duration::Weeks(52 * 5),
+        Instant::Now() - Duration::Weeks(52 * 4.9),
+        Duration::Days(7),
+        "NRLMSISE Past Data limit pass"
+    )
+                    // [TBI]: This test case fails sporadically because the actual data files are never updated based on
+                    // age, meaning they can get stale.
+
+                    // std::make_tuple(
+                    //     Instant::Now() + Duration::Weeks(52 * 0.9),
+                    //     Instant::Now() + Duration::Weeks(52 * 1.0),
+                    //     Duration::Days(7),
+                    //     "IERS Finals2000A Future Data limit pass"
+                    // )
+    )
+);
+
+TEST_P(OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Success, EdgeCases_Input_Data_Horizon)
+{
+    // long-horizon, low granularity propagation to test input data availability
+
+    // currently limited by:
+    // -5 years in the past due to CSSI space weather data when computing drag
+    // +1 year in the future due to finals2000a prediction span when computing polar motion for frame conversions
+
+    const auto parameters = GetParam();
+
+    const Instant startInstant = std::get<0>(parameters);
+    const Instant endInstant = std::get<1>(parameters);
+    const Duration step = std::get<2>(parameters);
+
+    const Real mass = satelliteDryMass_.inKilograms();
+    const Real surfaceArea = satelliteSystem_.getCrossSectionalSurfaceArea();
+    const Real dragCoefficient = satelliteSystem_.getDragCoefficient();
+
+    VectorXd initialStateVector(9);
+    initialStateVector << 7000000.0, 0.0, 0.0, 0.0, 5335.865450622126, 5335.865450622126, mass, surfaceArea,
+        dragCoefficient;
+
+    // Current state and instant setup
+    const State state = {startInstant, initialStateVector, gcrfSPtr_, dragCoordinatesBrokerSPtr_};
+
+    const NumericalSolver RK4 = {
+        NumericalSolver::LogType::NoLog,
+        NumericalSolver::StepperType::RungeKutta4,
+        step.inSeconds(),
+        1.0e-15,
+        1.0e-15,
+    };
+
+    // Setup dynamics
+    const Earth earth = Earth::FromModels(
+        std::make_shared<EarthGravitationalModel>(EarthGravitationalModel::Type::EGM2008),
+        std::make_shared<EarthMagneticModel>(EarthMagneticModel::Type::Undefined),
+        std::make_shared<EarthAtmosphericModel>(EarthAtmosphericModel::Type::NRLMSISE00)
+    );
+
+    const Shared<Celestial> earthSPtr = std::make_shared<Celestial>(earth);
+    const Array<Shared<Dynamics>> dynamics = {
+        std::make_shared<PositionDerivative>(),
+        std::make_shared<CentralBodyGravity>(earthSPtr),
+        std::make_shared<AtmosphericDrag>(earthSPtr),
+    };
+
+    // Setup Propagator model and orbit
+    const Propagator propagator = {
+        RK4,
+        dynamics,
+    };
+
+    // Propagate all states
+    EXPECT_NO_THROW(propagator.calculateStateAt(state, endInstant));
+}
+
+class OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Failure
+    : public OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation,
+      public ::testing::WithParamInterface<Tuple<Instant, Instant, Duration, std::string>>
+{
+};
+
+INSTANTIATE_TEST_SUITE_P(
+    Input_Data_Spans_Invalid,
+    OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Failure,
+    testing::Values(
+        std::make_tuple(
+            Instant::Now() - Duration::Weeks(52 * 6.2),
+            Instant::Now() - Duration::Weeks(52 * 6.3),
+            Duration::Days(7),
+            "NRLMSISE Past Data limit fail"
+        ),
+        std::make_tuple(
+            Instant::Now() + Duration::Weeks(52 * 1.0),
+            Instant::Now() + Duration::Weeks(52 * 1.1),
+            Duration::Days(7),
+            "IERS Finals2000A Future Data limit fail"
+        )
+    )
+);
+
+TEST_P(OpenSpaceToolkit_Astrodynamics_Validation_NRLMSIS00Validation_Data_Failure, EdgeCases_Input_Data_Horizon)
+{
+    // long-horizon, low granularity propagation to test input data availability
+
+    // currently limited by:
+    // -5 years in the past due to CSSI space weather data when computing drag
+    // +1 year in the future due to finals2000a prediction span when computing polar motion for frame conversions
+
+    auto parameters = GetParam();
+
+    Instant startInstant = std::get<0>(parameters);
+    Instant endInstant = std::get<1>(parameters);
+    Duration step = std::get<2>(parameters);
+
+    const Real mass = satelliteDryMass_.inKilograms();
+    const Real surfaceArea = satelliteSystem_.getCrossSectionalSurfaceArea();
+    const Real dragCoefficient = satelliteSystem_.getDragCoefficient();
+
+    VectorXd initialStateVector(9);
+    initialStateVector << 7000000.0, 0.0, 0.0, 0.0, 5335.865450622126, 5335.865450622126, mass, surfaceArea,
+        dragCoefficient;
+
+    // Current state and instant setup
+    const State state = {startInstant, initialStateVector, gcrfSPtr_, dragCoordinatesBrokerSPtr_};
+
+    const NumericalSolver RK4 = {
+        NumericalSolver::LogType::NoLog,
+        NumericalSolver::StepperType::RungeKutta4,
+        step.inSeconds(),
+        1.0e-15,
+        1.0e-15,
+    };
+
+    // Setup dynamics
+    const Earth earth = Earth::FromModels(
+        std::make_shared<EarthGravitationalModel>(EarthGravitationalModel::Type::EGM2008),
+        std::make_shared<EarthMagneticModel>(EarthMagneticModel::Type::Undefined),
+        std::make_shared<EarthAtmosphericModel>(EarthAtmosphericModel::Type::NRLMSISE00)
+    );
+
+    const Shared<Celestial> earthSPtr = std::make_shared<Celestial>(earth);
+    const Array<Shared<Dynamics>> dynamics = {
+        std::make_shared<PositionDerivative>(),
+        std::make_shared<CentralBodyGravity>(earthSPtr),
+        std::make_shared<AtmosphericDrag>(earthSPtr),
+    };
+
+    // Setup Propagator model and orbit
+    const Propagator propagator = {RK4, dynamics};
+
+    // Propagate all states
+    EXPECT_THROW(propagator.calculateStateAt(state, endInstant), ostk::core::error::RuntimeError);
+}
